pyproject.toml
[project]
name = "practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiofiles>=25.1.0",
    "aiolimiter>=1.2.1",
    "fsspec>=2026.4.0",
    "gcsfs>=2026.4.0",
    "httpx>=0.28.1",
    "prometheus-client>=0.25.0",
    "psycopg2-binary>=2.9.12",
    "python-dotenv>=1.2.2",
    "python-logging-loki>=0.3.1",
    "tenacity>=9.1.4",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["pipeline"]
